+++ ViAn/ViAn.pro
@@ -56,25 +56,25 @@ RESOURCES += resources.qrc
 # OPENCV
 #
 SOURCES += Video/video_player.cpp \
-    Video/shapes/Arrow.cpp \
-    Video/shapes/Circle.cpp \
-    Video/shapes/Line.cpp \
-    Video/shapes/Pen.cpp \
-    Video/shapes/Rectangle.cpp \
-    Video/shapes/Shape.cpp \
-    Video/shapes/Text.cpp \
-    Video/shapes/ZoomRectangle.cpp \
-    Video/Overlay.cpp
+    Video/overlay.cpp \
+    Video/shapes/arrow.cpp \
+    Video/shapes/circle.cpp \
+    Video/shapes/line.cpp \
+    Video/shapes/pen.cpp \
+    Video/shapes/rectangle.cpp \
+    Video/shapes/shape.cpp \
+    Video/shapes/text.cpp \
+    Video/shapes/zoomrectangle.cpp
 HEADERS += Video/video_player.h \
-    Video/shapes/Arrow.h \
-    Video/shapes/Circle.h \
-    Video/shapes/Line.h \
-    Video/shapes/Pen.h \
-    Video/shapes/Rectangle.h \
-    Video/shapes/Shape.h \
-    Video/shapes/Text.h \
-    Video/shapes/ZoomRectangle.h \
-    Video/Overlay.h
+    Video/overlay.h \
+    Video/shapes/arrow.h \
+    Video/shapes/circle.h \
+    Video/shapes/line.h \
+    Video/shapes/pen.h \
+    Video/shapes/rectangle.h \
+    Video/shapes/shape.h \
+    Video/shapes/text.h \
+    Video/shapes/zoomrectangle.h
 win32 {
     INCLUDEPATH += C:\opencv\release\install\include
     LIBS += C:\opencv\release\bin\libopencv_core320.dll
